Vision improvements
capture_and_analyze uses newest vision library functions as a test. vision library now uses a dummy position that sets off a drop-off alert when a dot is not seen, and the calibration method was fixed. The laser "library" now has a class that stores the laser pin and activates it